Extract worker workload contract
Move WorkloadClass into a provider-neutral worker module so task
architecture enforcement no longer depends on the Kombu topology module.
